Relocate SourceProvider-classes
