+++ test-common/src/test/java/io/github/chrimle/o2jrm/tests/GeneratedEnumTests.java
@@ -31,6 +31,7 @@ import io.github.chrimle.o2jrm.test.meta.configs.ConfigOption;
 import io.github.chrimle.o2jrm.test.meta.configs.Library;
 import io.github.chrimle.o2jrm.test.meta.filters.AssumptionFilter;
 import io.github.chrimle.o2jrm.test.meta.filters.Condition;
+import io.github.chrimle.o2jrm.test.meta.filters.GeneratedEnumProvider;
 import io.github.chrimle.o2jrm.test.meta.models.GeneratedField;
 import io.github.chrimle.o2jrm.test.meta.models.GeneratedSource;
 import io.github.chrimle.o2jrm.test.utils.CustomAssertions;
